+++ README.md
@@ -6,9 +6,9 @@
 <b>InitialForm</b> наследуется от <b>Form</b> - словоформа в начальной форме слова, имеет ссылки на все производные словоформы (<b>WordfForm</b>). <br>
 <b>MorphologyParameters</b> - набор констант для морфологических характеристик. <br>
 Пример работы с библиотекой описан в <b>Running</b> <br>
-<b>ВНИМАНИЕ РЕАЛИЗОВАН ФУНКЦИОНАЛ В РЕЖИМЕ АНАЛИЗА И ЧАСТИЧНО В РЕЖИМЕ ГЕНЕРАЦИИ! </b>
 <br>
-v 2.10 от 30.12.2017 <br>
+<br>
+v 2.10.1 от 30.12.2017 <br>
 1) Символьное представление словоформы теперь храниться в базе данных (для начальной формы в <b>dictionary.initialFormString.bd</b>, для производной формы в <b>dictionary.wordFormString.bd</b>, что уменьшает издержки в ОП. <br>
 2) База данных символьного представления и морфологические параметры хранятся в отдельные сборки MorphologicalStructures, что позволяет обрабатывать промежуточные результаты без JMorfSdk. <br>
 3) Добавлен режим генерации словоформе по начальной форме и наборы морф. характеристик. <br>
@@ -18,4 +18,9 @@ v 2.10 от 30.12.2017 <br>
 7) Все методы для работы с библиотекой возвращают Exception в случае, если слово не было найдено в словаре или случилось иная ошибка. <br>
 8) LoadJMorfSdk переименован в JMorfSdkLoad. <br>
 9) Если не нужна генерация слова, то можно удалить базу <b>dictionary.wordFormString.bd</b>, но тогда загружать библиотеку следует с помощью <b>JMorfSdkLoad.loadInAnalysisMode</b> <br>
+
+v 2.10.2 от 30.12.2017 <br>
+1. Перенес морф. параметров (MorphologyParameters) из JMorfSdk в проект MorphologicalStructures.<br>
+2. БД dictionary.initialFormString и dictionary.wordFormString, которые хранят стринговое представление начальной и производных формы слова, соответственно, перенесены в проект  MorphologicalStructures. Теперь сохранив ключ, полученный входе работы JMorfSdk, можно не хранить стринговое представление в ОП и получить его из БД (через проект MorphologicalStructures), когда она понадобится.<br>
+
